fix: move add event logic to services

+++ src/main/java/ru/yandex/practicum/filmorate/service/FilmService.java
@@ -2,16 +2,18 @@ package ru.yandex.practicum.filmorate.service;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Service;
 import ru.yandex.practicum.filmorate.controller.FilmController;
 import ru.yandex.practicum.filmorate.exception.ValidationException;
+import ru.yandex.practicum.filmorate.model.Event;
 import ru.yandex.practicum.filmorate.model.Film;
 import ru.yandex.practicum.filmorate.storage.director.DirectorStorage;
+import ru.yandex.practicum.filmorate.storage.event.EventStorage;
 import ru.yandex.practicum.filmorate.storage.film.FilmStorage;
 import ru.yandex.practicum.filmorate.storage.user.UserStorage;
 
+import java.time.Instant;
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 import java.util.Collection;
@@ -20,17 +22,27 @@ import java.util.stream.Collectors;
 
 @Service
 public class FilmService {
-	@Autowired
-	@Qualifier("FilmDbStorage")
-	private FilmStorage filmStorage;
-	@Autowired
-	@Qualifier("UserDbStorage")
-	private UserStorage userStorage;
-	@Autowired
-	@Qualifier("DirectorDbStorage")
-	private DirectorStorage directorStorage;
+	private final FilmStorage filmStorage;
 
-	private static final LocalDate startReleaseDate = LocalDate
+	private final UserStorage userStorage;
+
+	private final DirectorStorage directorStorage;
+
+    private final EventStorage eventStorage;
+
+    public FilmService(
+            @Qualifier("FilmDbStorage") FilmStorage filmStorage,
+            @Qualifier("UserDbStorage") UserStorage userStorage,
+            @Qualifier("DirectorDbStorage") DirectorStorage directorStorage,
+            EventStorage eventStorage
+    ) {
+        this.filmStorage = filmStorage;
+        this.userStorage = userStorage;
+        this.directorStorage = directorStorage;
+        this.eventStorage = eventStorage;
+    }
+
+    private static final LocalDate startReleaseDate = LocalDate
 			.parse("28.12.1895", DateTimeFormatter.ofPattern("dd.MM.yyyy"));
 
 	private static final Logger log = LoggerFactory.getLogger(FilmController.class);
@@ -79,12 +91,34 @@ public class FilmService {
 
 	public void addLike(Long filmId, Long userId) {
 		userStorage.getUser(userId);
-		filmStorage.addLike(filmId, userId);
+        Film film = getFilm(filmId);
+        if (!film.getLikes().contains(userId)) {
+            filmStorage.addLike(filmId, userId);
+            Event event = Event.builder()
+                    .userId(userId)
+                    .entityId(filmId)
+                    .timestamp(Instant.now().toEpochMilli())
+                    .eventType(Event.EventType.LIKE)
+                    .operation(Event.Operation.ADD)
+                    .build();
+            eventStorage.addEvent(event);
+        }
 	}
 
 	public void deleteLike(Long filmId, Long userId) {
 		userStorage.getUser(userId);
-		filmStorage.deleteLike(filmId, userId);
+        Film film = getFilm(filmId);
+        if (film.getLikes().contains(userId)) {
+            filmStorage.deleteLike(filmId, userId);
+            Event event = Event.builder()
+                    .userId(userId)
+                    .entityId(filmId)
+                    .timestamp(Instant.now().toEpochMilli())
+                    .eventType(Event.EventType.LIKE)
+                    .operation(Event.Operation.REMOVE)
+                    .build();
+            eventStorage.addEvent(event);
+        }
 	}
 
 	public Collection<Film> getSortedDirectorsFilms(Long directorId, String sortBy) {
